% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.plot.spm.R
\name{accuracy.plot.spm}
\alias{accuracy.plot.spm}
\title{Accuracy plot}
\usage{
accuracy.plot.spm(x, y, main, colramp, xbins = xbins., rng = "nat")
}
\arguments{
\item{x}{}

\item{y}{}

\item{main}{}

\item{colramp}{}

\item{xbins}{}

\item{rng}{}
}
\value{
plt


}
\description{
Accuracy plot
}
