% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_plot_spm.R
\name{accuracy_plot_spm}
\alias{accuracy_plot_spm}
\title{Accuracy plot}
\usage{
accuracy_plot_spm(x, y, main, colramp, xbins = xbins., rng = "nat")
}
\arguments{
\item{x}{observation vector}

\item{y}{estimated vector}

\item{main}{title}

\item{colramp}{color}

\item{xbins}{number of bins}

\item{rng}{methods for data representaion e.g., `norm`; normalized , `nat`;real values}
}
\value{
plt accuracy plot
}
\description{
Accuracy plot
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
