% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.tif.R
\name{extract.tif}
\alias{extract.tif}
\title{Extract points from a GeoTIFF using spatiotemporal overlay}
\usage{
extract.tif(
  df,
  date,
  tif,
  date.tif.begin,
  date.tif.end,
  coords = c("x", "y"),
  crs = "+init=epsg:3035"
)
}
\arguments{
\item{df}{data frame with coordinates and dates,}

\item{date}{date column in \code{df},}

\item{tif}{tif filename,}

\item{date.tif.begin}{reference begin date for the tif filename,}

\item{date.tif.end}{reference end date for the tif filename,}

\item{coords}{coordinate columns e.g. \code{"x", "y"},}

\item{crs}{projection system for the coordinates,}
}
\value{
data frame with point IDs and results of overlay
}
\description{
Extract points from a GeoTIFF using spatiotemporal overlay
}
\details{
Extends the \code{terra::extract} functionality. 
Works extremely fast if the GeoTIFF are Cloud-Optimized and located on SSD or similar.
}
\examples{
\dontrun{
library(terra)
library(dplyr)
library(parallel)
## run overlay in parallel:
tif.lst = list.files(pattern=".tif", full.names=TRUE) 
date.tif.lst = c(2000:2020)
cores = ifelse(parallel::detectCores()<length(tif.lst), parallel::detectCores(), length(tif.lst))
ov.pnts = parallel::mclapply(1:length(tif.lst), function(i){ 
  extract.tif(df, date="Year", tif=tif.lst[i], date.tif.begin=date.tif.lst[i]) }, 
  mc.cores = cores )
ov.pnts = do.call(rbind, ov.pnts)
str(ov.pnts)
}
}
