% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spm.R
\name{plot_spm}
\alias{plot_spm}
\title{Accuracy plot}
\usage{
plot_spm(
  df = NULL,
  main = NULL,
  palet = NULL,
  colorcut = NULL,
  xbins = 60,
  gvar_imp = TRUE,
  gtype = c("accuracy", "correlation", "var.imp"),
  gmode = c("root", "log10", "norm", "log2", "nat"),
  aspect = 1,
  ...
)
}
\arguments{
\item{df}{data frame}

\item{main}{}

\item{palet}{}

\item{colorcut}{}

\item{xbins}{}

\item{gvar_imp}{}

\item{gtype}{}

\item{gmode}{}

\item{aspect}{}

\item{...}{}
}
\value{

}
\description{
Accuracy plot
}
\examples{

}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
