% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spm.R
\name{predict_spm}
\alias{predict_spm}
\title{predict_spm}
\usage{
predict_spm(train.model, newdata, task = NULL)
}
\arguments{
\item{train.model}{trained model}

\item{newdata}{data at new location}
}
\value{
y response values
}
\description{
prediction on new dataset
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
