% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spm.R
\name{plot_spm}
\alias{plot_spm}
\title{Title}
\usage{
plot_spm(
  x,
  y,
  z = NULL,
  main = NULL,
  palet = NULL,
  colorcut = c(0, 0.01, 0.03, 0.07, 0.15, 0.25, 0.5, 0.75, 1),
  xbins = 60,
  gvar_imp = TRUE,
  gmode = c("root", "log10", "norm", "log2", "nat"),
  gtype = c("accuracy", "correlation"),
  aggregate = NULL,
  aspect = 1
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{z}{}

\item{main}{}

\item{palet}{}

\item{colorcut}{}

\item{xbins}{}

\item{gvar_imp}{}

\item{gmode}{}

\item{gtype}{}

\item{aggregate}{}

\item{aspect}{}
}
\value{

}
\description{
Title
}
