% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spm.R
\name{plot_spm}
\alias{plot_spm}
\title{Accuracy plot}
\usage{
plot_spm(
  df = NULL,
  main = NULL,
  palet = NULL,
  colorcut = NULL,
  xbins = 60,
  gvar_imp = TRUE,
  gtype = c("accuracy", "correlation", "var.imp"),
  gmode = c("root", "log10", "norm", "log2", "nat"),
  aspect = 1,
  ...
)
}
\arguments{
\item{df}{a data frame containing x,y, and z derived from `train_spm` and `predict_spm`,}

\item{main}{Title of the plot,}

\item{palet}{default values is palet=colorRampPalette(c("wheat2","yellow" ,"red","red3","orchid","orchid4),}

\item{colorcut}{default value is colorcut = c(0,0.01,0.03,0.07,0.15,0.25,0.5,0.75,1) ,}

\item{xbins}{number of bins, default value is 50,}

\item{gvar_imp}{variable importance,}

\item{gtype}{graphical type; user can choose among: gtype = c("accuracy", "correlation","var.imp")."accuracy" provides accuracy plot for a regression matrix. "correlation" provides partial correlation plot for the regression matrix. "var.imp" provides a graph of top 10 percent of the most informative features,}

\item{gmode}{graphical mode; in case `gtype=accuracy` gives user five options for representation of the accuracy plot as following: c("root","log10","norm","log2","nat"). "root" shows root square representation of the data,"norm" represent normalized representation of the data, "nat" represent natural  representation of the data,}

\item{aspect}{default values is aspect = 1,}

\item{...}{other arguments that can be passed on to \code{hexbin}.}
}
\description{
Dedicated function to plot diagnostic graphs resulted from `train_spm ` and `predict_spm`,
}
\examples{
\dontrun{
plt = eumap::plot_spm(df, gmode  = "norm" , gtype = "var.imp")
}
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
