% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.spm.R
\name{predict.spm}
\alias{predict.spm}
\title{predict.spm}
\usage{
\method{predict}{spm}(df.ts, task = NULL, train.model)
}
\arguments{
\item{df.ts}{}

\item{task}{}
}
\value{
y
}
\description{
prediction on new dataset
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
