% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spm.R
\name{predict_spm}
\alias{predict_spm}
\title{predict_spm}
\usage{
predict_spm(train_model, newdata)
}
\arguments{
\item{train_model}{trained model using `train_spm`,}

\item{newdata}{data at new location,}
}
\value{
predicted variables, variable importance
}
\description{
prediction on new dataset
}
\examples{
\dontrun{
predict.variable = eumap::predict_spm(train_model, newdata)
}
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
