% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spm.R
\name{predict_spm}
\alias{predict_spm}
\title{predict_spm}
\usage{
predict_spm(train_model, newdata)
}
\arguments{
\item{newdata}{data at new location}

\item{train.model}{trained model from `train_spm`}
}
\value{
y response values

vrimp variable importance value
}
\description{
prediction on new dataset
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
