% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spm.R
\name{predict_spm}
\alias{predict_spm}
\title{Title}
\usage{
predict_spm(train_model, newdata, task = NULL)
}
\arguments{
\item{train_model}{}

\item{newdata}{}

\item{task}{}
}
\value{

}
\description{
Title
}
