% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spm.R
\name{predict_spm}
\alias{predict_spm}
\title{Predict_spm}
\usage{
predict_spm(object, newdata)
}
\arguments{
\item{object}{Trained model using \code{train_spm},}

\item{newdata}{Data at new location,}
}
\value{
Predicted variables, variable importance
}
\description{
Prediction on a new dataset
}
\examples{
\dontrun{
predict.variable = eumap::predict_spm(object, newdata)
} 
}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa}
}
