% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_spm.R
\name{train_spm}
\alias{train_spm}
\title{Train Spatial matrix}
\usage{
train_spm(
  df.tr,
  target.variable,
  parallel = TRUE,
  predict_type = NULL,
  folds = 5,
  n_evals = 5,
  method.list = NULL,
  var.imp = NULL,
  super.learner = NULL,
  crs = NULL,
  coordinate_names = c("x", "y"),
  ...
)
}
\arguments{
\item{df.tr}{observation data}

\item{target.variable}{target.variable response variable}

\item{parallel}{parallel processing mode}

\item{predict_type}{e.g., response and prob}

\item{folds}{sub-item for spcv}

\item{n_evals}{number of evaluation process}

\item{method.list}{learning methods}

\item{var.imp}{variable importance}

\item{super.learner}{super learner}

\item{crs}{coordinate reference system, necessary for spcv}

\item{coordinate_names}{}

\item{...}{other arguments that can be passed on to \code{mlr3spatiotempcv::TaskSupervised},}
}
\value{
Object of class \code{mlr3}
}
\description{
This is a function to train (spatial) dataframe \herf{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0169748}{using Ensemble Machine Learning} and {https://mlr3.mlr-org.com/}{mlr3} ecosystem.
}
\examples{
## Meuse Demo
library(sp)
library(mlr3verse)
library(mlr3spatiotempcv)
library(checkmate)
library(future)
library(progressr)
library(eumap)
demo(meuse, echo=FALSE)
df <- as.data.frame(meuse)
df.grid <- as.data.frame(meuse.grid)
df = na.omit(df[,])
df.grid = na.omit(df.grid[,])
smp_size <- floor(0.8 * nrow(df))
set.seed(123)
train_ind <- sample(seq_len(nrow(df)), size = smp_size)
df.tr <- df[, c("x","y","dist","ffreq","soil","lead")]
df.ts <- df.grid[, c("x","y","dist","ffreq","soil")]
newdata <-df.ts
tr = eumap::train_spm(df.tr, target.variable = "lead", folds = 5 ,n_evals = 3,#' crs = "+init=epsg:3035")
train_model= tr[[1]]
var.imp = tr[[2]]
summary = tr[[3]]
response = tr[[4]]
vlp = tr[[5]]
target = tr[[6]]
predict.variable = eumap::predict_spm(train_model, newdata)
pred.v = predict.variable[[1]]
valu.imp= predict.variable[[2]]
plt = eumap::plot_spm(df, gmode  = "norm" , gtype = "var.imp")
df.ts$leadp = predict.variable
coordinates(df.ts) <- ~x+y
proj4string(df.ts) <- CRS("+init=epsg:28992")
gridded(df.ts) = TRUE # creat raster output
## regression grid 
make a map using ensemble machine learning with spatial cross validation for the predicted #' variables (*lead* in this case). 
plot(df.ts[,"leadp"])
points(meuse, pch="+")

}
\author{
\href{https://opengeohub.org/people/mohammadreza-sheykhmousa}{Mohammadreza Sheykhmousa} and  \href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
